# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText - OpenAPI 3.0
  description: |-
    This is the OpenAPI specification for the WASAText web-app.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /conversations:
    get:
      tags: ["conversations"]
      operationId: getMyConversations
      summary: Returns the conversations of the user
      description: |-
        Returns the lists of all the conversations of the user
      responses:
        "200":
          description: |-
            It is an array of integers
            containing the user's conversations' id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversations"
        "401":
          description: |-
            Unauthorized: missing or invalid bearer token
    post:
      tags: ["conversations"]
      operationId: createConversation
      summary: Create a new private conversation
      description: |-
        Creates a new private (one-to-one) conversation with the specified user
        and sends the first message at the same time.
        If a private conversation between the two users already exists, returns 409.
        The authenticated user is identified by the Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: |-
                  Request body for createConversation.
              type: object
              required:
                - targetUsername
                - content
              properties:
                targetUsername:
                  type: string
                  example: "Luigi"
                  minLength: 3
                  maxLength: 16
                  pattern: '^[A-Za-z0-9]{3,16}$'
                  description: |-
                      Username of the user to start a conversation with.
                content:
                  type: string
                  example: "Ciao"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 100000
                  description: |-
                      Text content of the first message.
      responses:
        "201":
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Bad request - Invalid username or message format
        "401":
          description: Unauthorized
        "404":
          description: Target user not found
          content:
            application/json:
              schema:
                description: |-
                  This schema contains the 404 error related message
                type: object
                properties:
                  error:
                    description: |-
                      The 404 error related message (user not found)
                    type: string
                    example: "User 'Luigi' not found"
        "409":
          description: Conflict - A private conversation with this user already exists
          content:
            application/json:
              schema:
                description: |-
                  This schema contains the 409 error related message
                type: object
                properties:
                  error:
                    description: |-
                      The 409 error related message
                    type: string
                    example: "A conversation with 'Luigi' already exists"
...
  /conversations/{conversationId}/messages/{messageId}/reactions:
    parameters:
      - name: conversationId
        in: path
        description: |-
            The id of the conversation
        required: true
        schema:
          type: integer
          minimum: 1
          example: 42
      - name: messageId
        in: path
        description: The id of the message to react to
        required: true
        schema:
          type: integer
          minimum: 1
          example: 18284
    put:
      tags: ["reactions"]
      operationId: commentMessage
      summary: Add or update reaction to a message (comment)
      description: |-
        Adds or updates a reaction (emoticon) to the specified message.
        Each user can have only ONE reaction per message.
        If the user already has a reaction on this message, it will be replaced,
        otherwise a new reaction is created.
        The user is identified by the Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionEmoticon"
      responses:
        "200":
          description: |-
              Reaction added or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionCreated"
        "400":
          description: |-
              Bad request - Invalid emoticon
        "401":
          description: |-
              Unauthorized
        "403":
          description: |-
              Forbidden - Not a participant in this conversation
        "404":
          description: Message not found
    delete:
      tags: ["reactions"]
      operationId: uncommentMessage
      summary: Remove reaction from message (uncomment)
      description: |-
        Removes the user's reaction (emoticon) from the specified message.
        The user is identified by the Bearer token.
      responses:
        "204":
          description: Reaction removed successfully
        "400":
          description: Bad request - Invalid message or conversation ID
        "401":
          description: Unauthorized
        "404":
          description: Message or reaction not found
...
  /groups:
    post:
      tags: ["groups"]
      operationId: createGroup
      summary: Create a new group
      description: |
        Creates a new group conversation with the specified members.
        The authenticated user becomes automatically a member and can invite others.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: |-
                  This schema represents the request body of the "createGroup" operation.
              type: object
              required:
                - name
                - memberUsernames
              properties:
                name:
                  type: string
                  example: "WASA"
                  minLength: 1
                  maxLength: 50
                  description: Group name
                  pattern:  '^[A-Za-z0-9]{3,100}$'
                memberUsernames:
                  type: array
                  minItems: 1
                  maxItems: 10000000
                  items:
                    type: string
                    minLength: 3
                    maxLength: 16
                    pattern:  '^[A-Za-z0-9]{3,100}$'
                  example: ["Luigi", "Marco", "Anna"]
                  description: Usernames of users to add to the group
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 201 response to the "createGroup" operation.
                type: object
                properties:
                  groupId:
                    description: |-
                        The id of the group created.
                    type: integer
                    example: 42
                  name:
                    description: |-
                        The name of the group.
                    type: string
                    example: "Gruppo Università"
                  members:
                    description: |-
                        The number of members of the group.
                    type: array
                    minItems: 1
                    maxItems: 10000000
                    items:
                      description: |-
                          Each item represents a username/member of the group
                      type: string
                    example: ["Maria", "Luigi", "Marco", "Anna"]
                  createdAt:
                    description: |-
                        Timestamp of the creation of the group.
                    type: string
                    format: date-time
        "400":
          description: Bad request - Invalid group data
        "401":
          description: Unauthorized
        "404":
          description: One or more users not found
...
  /groups/{groupId}/members:
    post:
      tags: ["groups"]
      operationId: addToGroup
      summary: Add a user to the group
      description: |-
        Adds a new member to the specified group.
        Only existing group members can add other users.
        The user making the request is identified by the Bearer token.
      parameters:
        - name: groupId
          in: path
          description: The id of the group
          required: true
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
      responses:
        "201":
          description: User added to group successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAddedToGroup"
        "400":
          description: Bad request - Invalid username
          content:
            application/json:
              schema:
                description: |-
                    The schema representing the 400 error of the "addToGroup" method.
                type: object
                properties:
                  error:
                    description: |-
                        A string representing the error.
                    type: string
                    example: "Invalid username format"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a member of this group
          content:
            application/json:
              schema:
                description: |-
                    The schema representing the 403 error of the "addToGroup" method.
                type: object
                properties:
                  error:
                    description: |-
                        A string representing the error.
                    type: string
                    example: "Only group members can add other users"
        "404":
          description: Group or user not found
          content:
            application/json:
              schema:
                description: |-
                    The schema representing the 404 error of the "addToGroup method".
                type: object
                properties:
                  error:
                    description: |-
                        A string representing the error.
                    type: string
                    example: "User 'Luigi' not found"
        "409":
          description: Conflict - User is already a member
          content:
            application/json:
              schema:
                description: |-
                    The error 409 of the "addToGroup" method.
                type: object
                properties:
                  error:
                    description: |-
                        A string representing the error.
                    type: string
                    example: "User 'Luigi' is already a member of this group"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        User identifier returned from login.
        Use the identifier as the bearer token.
  schemas:
    Conversations:
      description: |-
          This schema represents a list of conversations
      type: array
      minItems: 0
      maxItems: 100000
      items:
        type: object
        properties:
          id:
            type: integer
            minimum: 1
            example: 42
            description: |-
                Conversation identifier
          type:
            type: string
            enum: [direct, group]
            example: direct
            description: |-
                Type of conversation (direct chat or group)
          name:
            type: string
            example: "Marco"
            description: |-
                Username (for direct) or group name (for group).
          photoUrl:
            type: string
            format: uri
            example: "https://example.com/photos/user123.jpg"
            description: |-
                Profile photo URL (user or group)
          lastMessage:
            type: object
            description: |-
                Preview of the last message
            properties:
              preview:
                type: string
                example: "Ciao, come stai?"
                maxLength: 100
                description: |-
                    Text snippet (photo url if photo message)
              timestamp:
                type: string
                format: date-time
                example: "2025-12-13T17:37:43Z"
                description: |-
                    Timestamp of last message
              isPhoto:
                type: boolean
                example: false
                description: |-
                    True if last message is a photo
          unreadCount:
            type: integer
            minimum: 0
            example: 3
            description: |-
                Number of unread messages in this conversation
    Conversation:
      description: |-
          This schema represents a conversation.
      type: object 
      properties:
        id:
          type: integer
          minimum: 1
          description: |-
              The conversation identifier
        type:
          type: string
          enum: [direct, group]
          example: direct
          description: |-
              Type of conversation
        name:
          type: string
          example: "Marco"
          description: |-
              Username or group name
        photoUrl:
          type: string
          format: uri
          example: "https://example.com/photos/user123.jpg"
          description: |-
              Profile photo URL
        participants:
          type: array
          description: |-
              List of participant usernames (for groups)
          minItems: 1
          maxItems: 10000000
          items:
            description: |-
                Each item of the array represents a single partecipant of the conversation.
            type: string
            example: "Maria"
        messages:
          type: array
          description: |-
              All messages in reverse chronological order
          minItems: 0
          maxItems: 100000000
          items:
            type: object
            properties:
              id:
                description: |-
                    The id of the message
                type: integer
                minimum: 1
                example: 18284
              senderId:
                type: string
                example: "user123"
                description: |-
                    ID of the message sender
              senderUsername:
                type: string
                example: "Maria"
                description: |-
                    Username of sender
              content:
                type: string
                example: "Ciao, come stai?"
                description: |-
                    Text content (empty if photo message)
              photoUrl:
                type: string
                format: uri
                example: "https://example.com/photos/msg123.jpg"
                description: |-
                    Photo URL (present only for photo messages)
              timestamp:
                description: |-
                    The timesatmp of the message.
                type: string
                format: date-time
                example: "2025-12-13T17:37:43Z"
              status:
                type: string
                enum: [sent, received, read]
                example: "received"
                description: |-
                  Message status (checkmarks - for sent messages):
                    - sent: message sent (no checkmark)
                    - received: one checkmark (delivered to all recipients)
                    - read: two checkmarks (opened by all recipients)
              replyTo:
                type: object
                description: |-
                    Present only if this message is a reply
                properties:
                  messageId:
                    description: |-
                        The id of the message to be replied to.
                    type: integer
                    example: 18200
                  content:
                    description: |-
                        The content of the message to be replied to.
                    type: string
                    example: "Ci sei?"
                  senderUsername:
                    description: |-
                        The username of user who sent the message to be replied to
                    type: string
                    example: "Luigi"
              reactions:
                type: array
                description: |-
                    Comments/reactions on this message
                minItems: 0
                maxItems: 100000
                items:
                  type: object
                  properties:
                    username:
                      description: |-
                          The username of the user who reacted.
                      type: string
                      example: "Marco"
                    emoticon:
                      type: string
                      example: "👍"
                      description: |-
                          Reaction emoticon
                    timestamp:
                      description: |-
                          The timestamp of the reaction.
                      type: string
                      format: date-time
                      example: "2025-12-13T17:38:00Z"
              isForwarded:
                type: boolean
                example: false
                description: |-
                    True if this message was forwarded from another conversation
...
    ReactionEmoticon:
      description: |-
          This schema represents the emoticon of a reaction.
      type: object
      required:
        - emoticon
      properties:
        emoticon:
          type: string
          pattern:  '^.*$'
          example:  "😄"
          minLength: 1
          maxLength: 10
          description: |-
              Emoticon to use as reaction (emoji or text)
    ReactionCreated:
      description: |-
          A reaction just created.
      type: object
      properties:
        username:
          description: |-
              The username of the user who sent a reaction.
          type: string
          example: "Maria"
        emoticon:
          description: |-
              The emoticon of the reaction.
          type: string
          example: "😄"
        messageId:
          description: |-
              The id of the message of the reaction.
          type: integer
          example: 18284
...
    Username:
      description: |-
          This schema represents a username that has to be created(it is supposed to be used for client's requests).
      type: object
      required:
        - username
      properties:
        username:
          type: string
          example: "NewMaria"
          minLength: 3
          maxLength: 16
          pattern: '^[A-Za-z0-9]{3,16}$'
          description: |-
              New username (must be unique)
...
    UsernameAddedToGroup:
      description: |-
          This schema represents a user added to a group. Hence it is characterized by 3 properties: groupId, username, addedAt(timestamp).
      type: object
      properties:
        groupId:
          description: |-
              The id of the group.
          type: integer
          example: 42
        username:
          description: |-
              The username of the user added to the group.
          type: string
          example: "Luigi"
        addedAt:
          description: |-
              The timestamp of the operation.
          type: string
          format: date-time
          example: "2025-12-13T18:00:00Z"
security:
  - bearerAuth: []